remove_smatrix.py
